refactor(mind-map): extract fn to index.ts
